requirements-web.txt
onnx
onnxruntime
optimum-onnxruntime
